[002] remove VirtualCameraManager
