--- Assets/Project/Scripts/Managers/Core/VirtualCameraManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cbce012407cfc3c4aba3a0ee5225a50f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
